requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Vector store
chromadb>=0.4.18
# or
# pinecone-client>=2.2.4
# or
# qdrant-client>=1.7.0

# Search
elasticsearch>=8.11.0

# LLM（多提供商：OpenAI / DeepSeek / Claude / Qianwen）
openai>=1.3.0
anthropic>=0.18.0

# HTTP client
httpx>=0.25.2
aiohttp>=3.9.1

# Observability
# 固定版本以避免依赖冲突，确保与 Python 3.13 兼容
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.42b0,<1.0.0
opentelemetry-exporter-jaeger>=1.21.0,<2.0.0
opentelemetry-proto>=1.15.0,<2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2  # for testing

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
